Split out JS module
Also:
 - Neaten up some views
