+++ app/views/cards/_cards.html.erb
@@ -1,0 +1,5 @@
+<ul id="cards">
+  <% cards.each do |card| %>
+    <li><%= render card %></li>
+  <% end %>
+</ul>
